Adapt unit tests interfacing with rust to the API changes in the new version
The changes to the API were singificant causing a lots of noise and
restructuring in these tests.
Some of them are not valid anymore (dropped) or not working due to
some behaviors being broken in the library (planned to be addressed by
upcoming releases) or due to missing implementation on our side.
They are either ignored or commented out (shame on me) to be
easy to track and will be addressed by the next commit.

ET-1031