Bump rust library to 0.11.19 and adapt to
- Bump rust to 0.11.19
- Update conversation domain model to hold information that were before part
of "context" (eg. time, size, number of unread messages, number of attachments).
This is because rust now handles "context" for the clients, returning only
the right data for each label that is requested.
As a consequence, the `labels` field of Conversation model doesn't make sense
anymore and will be replaced by the newly added `customLabels` instead
(another field to hold `exclusiveLocation` will also be added)
- Labels are now exposed with dedicated types for System, CustomLabel and
CustomFolder. Updated mappers to be able to map them back to a unified list of
lables (allowing to reuse existing presentation logic);
- Use `allCustomFolders` method to read a flat hierarchy of custom folders
instead of one resolved to parent / children relationships
(allows to reuse presentation logic)
- Updated unread counters repository to build Counters object by merging
the now-differentiated list of labels (system, custom labels, folders)
- Fix all test and supporting test data

ET-1031