Fix switching locations not loading new conversations
This is kind of a hack: the combination of `destroy()` setting the
(mutable) Flow value to `null` and the returned flow filtering null
values, allows for the flow to "wait" till the new Mailbox is created
and the intial value emitted, which in turns updates the UI after the
user switched location.

There is still an issue in that while this happens, the user is left
looking at the previous location content while already in the new
location, till the emission actually happens. This is expected to change
heavily with the introduction of custom-pagination-solution from rust,
so deemed not worth to be tackled now.

Side note, Mailbox creation seems to be taking several seconds - added
in some verbose logs to troubleshoot it.

ET-1080