Add custom lables to Message domain model
Mapping it's value from rust and adjusting the higher level mapping to
MailboxItem to use this instead of running a computation using labelIds.
The major difference is that custom labels only contains custom lables
while labelIds did contain both custom and system ones.
Usages that were relying on system ones will be updated to use the
"exclusiveLocation" field that'll be added to message when exposed by rust.

This allows to streamline the logic to getMailboxItems as there's no
need to get labels at all.

ET-1063