Change signature of ObserveConversationMessages use case to return ConversationMessages
instead of ConversationMessagesWithLabels as before. All the methods in
the hierarchy (above and below this use case) have been updated to
reflect this change too.
The reason to do it is that "custom labels" are now provided as part of
the Message domain model (computed by the rust library and mapped to domain)
which removes the need for domain models to keep track of the whole
`labelIds` list (which was also removed in this commit for message) and
to run computations on it to define bits such as custom labels, starred
status of exclusive location (which will be exposed on the domain model
later).

Lots of areas of the code were relying on these constructs: they were
updated to use the new ones were meaningful, while functionalities that
we already know will shift to rust were treated more drastically
(mostly, dropped pointing them out with a @MissingRustApi annotation to
later introduce rust logic; Tests on legacy logic were also dropped
along with it). There functionalities include:
- Defining Message Location Icons
- Pre-selections of labels on the "label as" bottomsheet
- Hiding trash/spam messages of a conversation when opening it from
  locations other than trash/spam
- Draft icon on avatar when a message is a draft
- Setting "to list" on a Reply message based on the message.isSent
  status (ie. send to yourself if replying to a sent message)

In general, from here forward, anything relying on labelIds should be
considered unreliable (or plain wrong).

ET-1063