Import Label models from core project and replace all usages
This moves Label, LabelId and LabelType domain models from core and
replaces all usages. It also drops the dependency on core's label module.
The rationale behind this change is to allow updating the label model to
support rust-driven needs (eg. `color` field optionality) without having
a dependency on core (which for the most part is now deprecated).

ET-1140