Update Rust Message query to use paginator interface instead of message watcher
This basic usage should yield the same outcome as the previous
implementation, which is, emit one single page (the first one).
It will be updated afterwards to require different pages depending on
the given page key.

Note! As of lib 0.11.43 the onupdate callback is not working properly
(being called many times with no actual change happening). It was
disabled in this commit to keep things working.

ET-201