Do not observe view mode by location to create pager
This change fixes one issue and breaks a functionality:
- Issue fixed: given conversation mode enabled, when switching location the
"viewMode" (combine flow below) would emit before the
"selectedMailLabelId" causing the chain of call to pager > pagingSource
> Getmailbox > Repo to trigger and effectively loading the previously
selected label with the current one's viewMode. This only affects
ConversationMode on (in which mode we force message mode for certain
locations) as in message mode the view mode would never change.
- Broken functionality: without observeViewModeByLocation() we don't
  automatically re-create the pager when the user changes view mode
  in the settings. It's not clear how this will work thorugh the rust
  lib anyways, so this is considered to be a worthy tradeoff.

ET-201