Paginate message items from rust library
RustMessageDataSource defaults to empty list when query returns null
(error interface from rust lib is still to be defined, the current
expectations is for null to never be returned)

Refactor RustMessageQuery.getMessages to be a suspending function which
returns a list of message instead of a flow. This makes more sense for
the use case at hand (ie. sync loading of one page at a time) and fixes
a concurrency issue for which the data source would return the previous
value from the flow (causing the same page to be returned twice).

ET-201