Drop invalid fields from domain models due to the rust lib's structure
- Avoid passing userId data from UI to VM when performing move actions
  on mailbox (since userId is not mapped to domain>uiModels anymore);
  Read it from the primaryUser in VM instead;

- Drop AddressId and UserAddress from Message and MessageBody domain
  models (same reason, not exposed by rust thus can't be mapped);

- !Open in calendar feature broken!
Given AddressId not being exposed through with rust Message
(for both Message and Participants) resulting in the client not having enough
info to determine with "recipient" is the current user for which to open the invite.
Currently getting "toRecipients.first()" to keep the API unchanged,
this won't work in several cases.

ET-201