Fix invalidation observer being called as many times as loaded pages
Due to the ObserverWrapper (implementation detail of
RustInvalidationTracker) holding an instance of RustInvalidationObserver
rather than the Weak one, the call to onInvalidate was skipping the
login to track observers references weakly that's part of WeakObserver.

After a closer look to this logic, it appears we actually always want to
only have one observer (which is, the one for the current page) and that
it doesn't need to be weak (as it did previously due to being attached
to the room table) as it'll just be replaced by the next one as the
pages load. This lead to the simplification in this commit.

ET-201