Get Move To Bottom Sheet Actions use case returns valid locations to move to
This use case queries from the rust lib the available system locations
where the given list of messages OR conversations can be moved to
(should be context dependant) and separetely, gets the list of all
custom folders from label repository merges the two and returns them.

The result is that the "move to" available locations are all the custom
folders and a choosen list of system ones.
While the rust lib also exposes custom folders as part of the
"MoveToActions", we decided to get them from the existing label
repository (as `Label` domain objects) as this allows for easy mapping
to MailLabels which in turn maps to MailLabelUiModel and renders on the
UI considering hierarchy of folders etc. which would have had to be
duplicated if we were to map rust's MoveAction.Custom to MailLabels.

ET-1243