Show more actions bottom sheet for conversation
The sheet is filled with the "available actions" returned by the rust
lib for this conversation.
Some of them do not make sense on the conversation level (eg. reply,
viewInLightMode etc) and are being outright ignored (at the "click
listener" level, clicking them when the bottom sheet has no message id
fires nothing) and are being discussed.

ET-1243