Introduce HeaderValue + Converters
Since the value of the header is only known at runtime and once the value has been fetched, the app needs to handle both JsonLiterals and generic JsonObjects.
A header could then be a Map<String, String> or a Map<String, JsonObject>. The former implementation of the headers assumes that headers can only be Map<String, String>, so in case we get something different, the message fails to open due to a serialization error.

MAILANDR-1531