Allow second level bottom sheets to be shown
This fixes an issue for which requiring the "move to" or "label as"
bottomsheets through the "more actions" one would cause the latter to be
dismissed withouth the former to be shown.

This seems to be due to this "onDispose" being called even though it didn't
really leave the composition. Since the initial intention behind this disposition
isn't clear and we're quite few versions of compose after the one where it
was introduced, the tentative solution is to drop it.

ET-1345