Overload move conversations method to encapsulate resolution of local label id
when moving to system locations, we often use SystemLabelId.{Location} as
a mean to statically refer to the location we're moving to.
Given system label ids are now dynamically defined by rust (locally) and
change every session, this doesn't work without first mapping between the static
and the local id.
This allows to do such mapping in a single place.

ET-1345