Do not manually disconnect watcher for conversation detail
With the latest changes on rust side, now watchers are handled as follows:
> Your runtime destroys the WatchHandler
> We detect that it no longer exists
> We call disconnect
> We run the destructor of the callback you provided and our data structures

meaning there's no actual need to manually call disconnect when removing one
handle (as clearning the reference should automatically do it).

The trigger for this change was that we're not able to test the calls to
disconnect anymore due to the handle not existing at all (when mocking the
watcher).

ET-1421