Mailbox shows loading instead of previous items when switching to a new location
Before, the previous location's items were shown till the new paging data
is returned through the paging source (meaning, after it's loaded from
network).

The cachedIn(viewModelScope) that was previously applied to pager.flow in
mapPagingData function was dropped, as it sometimes causes a glitch where the
previous location's items are shown for a split second after the loading
and before the new ones.
It is not supposed to add any functionlity either, considering that the class
level `items` field is being cached in the same fashion.

ET-1647