Update ComposerViewModel to rely on rust-provided init logic
This commit comments out all of the code that is now delegated to rust init
methods, putting calls to the latter in place while trying to preserve
the existing events in place.
The commented bits are there to help rust devs get all the behavior in place
by comparison, and will be dropped in a standalone commit.

ET-1679

Drop all functionality that is now delegate to rust lib from Composer VM

ET-1679