Fix search functionality by delegating selection of label to search on to rust
Search always needs to happen in "AllMail"; Depending on user's preference,
this may also be "AlmostAllMail" (excluding trash / spam), which in turn
causes the manual lookup for "local label id" that we were running here
fail (as in the latter case, "AllMail" label is not available through the
rust lib).
With this change, we delegate to rust mailbox object the decision of which
of the two "all mail" to switch to, which restores search functionality.

ET-1689