Merging paginators for search and messages into one
This change addresses some issues caused by the cohesistence of two instances
of rust's "Paginator", due to the "message" paginator and the "search" one
being handled separately.
The most visible issue (while not the only) was that when switching to search
the message paginator (now in the background but still existing) would get
its "onUpdate" callback fired every second (or so), causing it to trigger an
invalidation that in turn would cause the (now active) search paginator to
reload - entering an inifinite load loop.

We attempted addressing this by making the involved classes non-singleton and
thus collectible by GC (on the working assumption this would kill the paginator
in the background and allow the new one to thrive) - didn't work out (DI
tree is relying on singletong till its roots, making it hard to bind some parts
to VM; didn't look like a fully reliable fix;)

ET-1689