Create ad hoc instances of Rust Mailbox when needed rather than sharing one
The main reason behind this commit is to start solving issues due to the shared
Mailbox object not being in the right state.
One example, is with search functionality: once we enter search, if opening any
results, we expect the "Mailbox" object to be holding a "AllMail" label (given
search always happens in all mail); instead, it contains the labelId of the
location we started the search from (eg. Inbox) as it's not updated by search;
If updating it, when existing search we would need to update it back.

What we're moving towards instead, is to have the UI pass through the labelId
(where needed) that the data layer will use to create a [rust] Mailbox object
anew (with some local caching if needed...) and not hold a reference to it.

Most notable changes in this commit:
- RustMailbox object dropped
- RustConversationDetailQueryImpl adds a field to store the given userId
to use it in the callback; Not cool.
- Do not set "Mailbox" values when creating new paginators (as paginator interface
is already working with labels, hence this step was there as a "preparatory" step
for when we get to the detail);
- Change labelId param to get messageBody to not be optional: this is to keep
things simple while changing it; The labelId is not really needed to
get the body and will be dropped on the rust side from this interface;
- Add a rustMailboxFactory.create(userId) method that allows creating without
passing a label; this is to allow reproducing the behavior we were allowing
before (ie. get a mailbox without having a label) and is implemented through
getting the "selectedLabel" from the Singleton class. This should be dropped
as we pass labelIds through to all methods;

ET-1689