Remove "current label id" parameter from get Message Body API
Clients are currently passing "current label id" info on some APIs where
it's not really needed (ie. to get message body for the case at hand);
This commit removes this info from being passed around in the client;
Since the rust API is still requiring it, we pass an Hardcoded "AllMail"
value as it shouldn't be used anyways and it is scheduled to be removed later on.

ET-1689