Hardcode "current label id" parameter for message actions
Some message actions (namely mark read, mark unread, and delete) are
currently passing to the corresponding rust API a "Mailbox" object, which has
the only purpose of passing the "current label id" (as is, in which mailbox
is this action being performed into).
Such information is not really needed to perform these actions and the rust
lib already own the data needed to perform all realated tasks (eg. update
unread counters etc).

Here we hardcode it to AllMail with the goal of removing the rustMailboxFactory.create(userId)
method that is error prone due to being basing the definion of "current label id"
on shared state (selectedLabelId) which isn't considering all cases (eg. search).

It should be removed from the rust lib soon.

ET-1689