Conversation detail View Model passes "context" label Id to open conversation
This change is part of a wider refactoring where we are stopping to hold
"current label id" (aka. "context", "opened from", "selected label id") in
shared, static objects (eg. "selectedMailLabel or rustMailbox)
to pass it along to ensure consistently of data.

The chain of calls to observeConversation and observeConversationMessages now
gets a labelId as param; this is meant to be the "context" we want to observe
the conversation in, which usually corresponds to the mailbox we open the
conversation from (eg. inbox, archive..).

Non straightforward changes:
- ConversationViewModel now expects to always have a "openedFrom" param
- NotificationsDeepLinksViewModel infers the label based on the received message
or falls back to all mail (by resolving it to a local label id for rust to be
able to interpret it);

ET-1689