Mailbox passes "opened from label id" to conversation detail considering search
When in search, the "context" (or "opened from") label Id should always be `AllMail`.
Due to the "exclude spam and trash from all mail" setting, it can also be
AlmostAllMail.
Clients do not directly deal with this setting, this commit
relies on the labelId that's made available through rust (which is either AllMail
or AlmostAllMail) to infer the context.

Note that following this setting, when in AlmostAllMail mode search should
also not show results from spam and trash (which wouldn't load given the
context was inferred to AlmostAllMail);
This hack will not suffice when adding the "show results from spam / trash")
feature. Solutions are being evaluated.

ET-1689