Find local system label id runs custom logic for All Mail or Almost All Mail
This addresses the issue detailed in 799dcc7 where trying to resolve AllMail
or Almost All Mail would fail as rust returns only one of the two based on
user settings (which clients do not have access to).
The new use case delegates to rust the resolution of all mail locations.

- Removed fallback between all mail and almost all mail from mailbox VM

ET-1689