Adapt all rust APIs to new rust error handling approach
Change all of the APIs interfacing with rust to handle the new return
type Result<Error, Ok>.

- In most cases, we mapped such result to either in the "Wrapper"
objects and did minor adaptations to the outer layers;
- In others we propagated the either where before optionals were used
- Overall, where error handling wasn't in place before on the client's side
(eg. handling optionals and doing nothing when empty)
the behavior was left the same (ie. didn't handle such errors
across the whole client).

ET-1748

- Adapt Observe session / account use cases to new rust error handling

- Adapt contact data source to new rust error handling approach

This assumes in the Data Source test we're able to mock rust's error types,
which is not a given (and maybe, not a good idea anyways).
Can't run the tests now to confirm as too many other areas need fixing,
will re-eval this once we get the app running again.

- Adapt label data source to new rust error handling approach

- Add basic mappings between rust and domain error types

- Adapt message data source to new rust error handling approach

- Adapt Mailbox creation to new rust error handling approach

- Adapt paginator to rust's new error handling approach

- Adapt get available message actions to new rust error handling

- Adapt get available message actions to rust new error handling

- Adapt get conversation message to rust new error handling

- Adapt get rust message label as actions to new error handling approach

- Adapt rust message get move to actions to new error handling approach

- Remove uneeded label Id param from get Message Body API

Same of dfd6376a4 in a less refined way (will reuse the same when rebased)

- Adapt Message Body API to new rust error handling approach

- Refactor MessageRepo > MessageDataSource to push handling of rust-related
objects from the repo to the data source (ie. DecryptedMessageWrapper operations
are now only in the DS).
- Adapt other usages of getLocalMessageWithBody to deal with the new "either"
return type in the easiest way to limit changes (ie. getOrNull and keep in place
the old optionality checks - this is fine as such usages are deprecated and need
rework to be used).

- Adapt rust Mail Session creation to the new error handling approach

- Adapt rust get all mail label Id to new error handling approach

- Adapt user session repo to rust new error handling approach

- Adapt user session fork to rust new error handling approach

- Adapt Login View Model to rust new error handling approach

- Adapt shared Password input view models to new rust error handling

- Adapt mail settings to new rust error handling approach

- Adapt draft API to new rust error handling approach

- Adapt conversation paginator interface to rust new error handling approach

- Adapt conversation data source to new rust error handling approach

- Adapt Observe Primary Core Account to rust new error handling

- Remove unneeded kapt correctErrorTypes from common build gradle