Bump rust lib to 0.44.0
This PR brings the project back to compiling by integrating the API changes.

There's one change that's pending to be checked, on the MessagePaginatorWrapper
and ConversationPaginatorWrapper now we get back either `Append` or `Replace`
objects which contain the list of items to display.
The expectation is that when receiving a `Replace`, all the list is replaced
by the new objects.

ET-2190