Composer VM builds a Draft Display Body and adds it to the state
This state will be used to show the draft body within the HTML template,
with injected css and javascript to allow working within the webview.

It needs to be a separated fields from the previously existing body as the
former will only be shown once when opening the composer and then not updated
(excluding some cases such as change sender which will be handled separately)
as the webview component is stateful and will hold its own state.

The latter (composerFields.body), will instad be updated to reflect the part
of the body that the user is editing (hence, without the template, css and js)
and will be used to store state to the rust lib.

ET-2208