Keep cursor in the visible view port while typing in composer
With the introduction of a webview to edit html, we lost some control over
some of the view's state (eg. scrolling); as a result, as a user types in new
lines the cursor can end up behind the IME and let the user typing blindly.

This solution emits an event from the webview when the body is resized (ie.
user typing in new lines), measures the size of the whole webview at that point
and passes that up to the composable hierarchy to allow adjusting the
scroll position accordingly.

ET-2208