Correct page loading after location changes
Rust paginator only provides nextPage and reload. Even we request the
first page from Mailbox UI, at the data layer we call nextPage. When a
location is first opened it works fine but we change to a different location
and return to the orginal location, Mailbox UI requests the firstPage but
we call nextPage from Rust paginator, and it returns a page of older messages.

Fix: If we already have a paginator and UI requests the firstPage then we
recreate the paginator.

ET-2449