Composer editor automatically brings cursor in focus when entering new lines
At this stage, cursor is only brought to focus when new lines are added to the
composer (either through typing more than one line or hitting enter);
That's due to the webview only emitting a "editorSizeChanged" event when
its height is updated. This can be changed to emit on character typed as well,
once that event reaches the scroll manager same logic should apply.

ET-2344