Group composer editor javascript code in single place
I initially evaluated extracting js to a separated file to keep things more
readable; I didn't go for it as we loose the possibility to reference kotlin
constants (eg. EDITOR_ID or JAVASCRIPT_CALLBACK_INTERFACE_NAME) which make the
whole thing arguably more error prone that not dealing with js as a string.

ET-2344