Drop verbose logs from rust label data source
This method can be called compulsively when mapping mailbox items to ui Models
(`MailboxItemUiModelMapper`) as it's part of the logic to check whether a
"location icon" should be displayed for the current "selected mail label".

In terms of performances this isn't a big deal as the local data source is
caching such data in memory; it is instead an issue in terms of pulluting the
logs with not-so-useful entries.

NOJIRA