Fix crash when switching user due to key reuse
When switching between two users where both have custom labels and folders,
due to the fact that SidebarCustomLabel adds both labels and folders to the
same lazylist, a crash can happen if the id of one of the labels
of user A collides with the id of one of the folders of user B.

That's due to the fact that at the time folders are added, the previous user's
labels are still composed (in the same LazyListScope) and thus adding user B's
one will complain the "key" is already existing.

This is fixed by adding the "type" of the folder to the key.

ET-2421