Fix Rust contact data source tests following watcher lifecycle update
Commit `66fe5844` updated the way this data source handles the lifecycle of
the "watcher", moving from a single-instance watcher kept alive throughout
the app lifecycle (given this class is injected as a Singleton) to
creating a new callbackFlow for each "observe" method invokation with explicit
handling of channel closure when not observed anymore.

Tests were updated to reflect such changes as well as to ensure that we still
emit errors when things go wrong.

ET-2386