Avoid long mailbox loading times when coming back from detail screen
As mentioned on commit 770c36 in regards to LabelRepository.observeSystemLabels:

> This method can be called compulsively when mapping mailbox items to ui Models
> (`MailboxItemUiModelMapper`) as it's part of the logic to check whether a
> "location icon" should be displayed for the current "selected mail label".

While this wasn't as issue before the changes in 52d1509, now on the
Data source we create a new watcher for each call making this expensive and
in turn causing several seconds of wait time when going back to mailbox.

This commit adds a cache layer on ShouldShowLocationIndicator, which
being part of the MailboxVM dependency tree will be kept in memory for the
whole lifecycle of Mailbox and ensure we only get system labels once for each
user.

ET-2386