Refresh attachments after pgp/mime message is decrypted
After the message body is decrypted through the normal flow, in case it's pgp/mime
rust will expose the decrypted mime attachments to the message "attachments" field.
Due to this we need to ask again the message to rust in order to ensure the
attachments being returned are complete.

This can be improved by full removing the attachments field from the body and
relying on attachments coming through with the "conversation messages" flow;
Rust would then ensure after decryption of the pgp mime body the flow emits
the updated attachments.

NOJIRA