Add Data Source impl to interact with rust contact suggestions API
This introduces a significant tradeoff as to allow reusing "ContactMetadata" model,
some abomination is done to map rust's "ContactSuggestions" models to it.
In particular, when we are mapping contact groups members the client side
expects more data than rust exposes (eg. name, id, avatarInfo);
The piece of data that really matters is email, thus we force the mapping
by using email in all fields that are missing.

Using a new model was also spiked, it's feasible but introduces lots of duplication
with the current one. Given this flow should be quite standalone (ie. data
that's missing isn't expected to be used).

ET-454