Avoid flickering when selecting a contact suggestion
The flickering effect is related to the show / hide logic in RecipientFields
composable (ie. hidden = suggestionField == null || suggestions.isEmpty())
and the fact that the "closeSuggestion" op sets suggestionField to null
(hiding them) and then, ChipsListField emits a "onSuggestionTermTyped" due
to the clear of the search field and this last op sets the suggestionField to
TO again (causing the show);

Removing the debouce on the contacts suggestions fixes this; Being querying
the suggestions locally from the rust lib this shouldn't have much impact.

ET-2161