Restore contact suggestions scrolling below "to", "cc" and "bcc" fields
This was dropped during conflicts resolution due to the previous patch not applying
cleanly. A new solution was applied, which calculates manually the size of the
shown "recipients" to take into account the hiding logic (ie. when typing in "to"
then "cc" and "bcc" are hidden etc) and the fact that suggestions are now part
of the field itself.

ET-2161