Do not show the message sent snackbar again when switching accounts
The RustSendingStatusDataSourceImpl is injected as a Singleton, thus was never
re-created when switching accounts but rather was re-emitting from the same flow,
meaning that after sending a message, any account switch would bring up the
"Message Sent" snack again.
With this commit, we introduce a callback flow that stops being observed when we
switch account.

ET-2648