Fix error with duplicated recipient when replying to a message
Due to the fact we don't yet pass recipients "name" thorugh from rust to UI
and backwards, it can happen (notabily, when replying to a message) that rust
initializes a draft with a "recipient" (with name and address) and the UI,
after initialization, tries to update recipients passing another "recipient"
with the same address but no name;
In this case, rust would return a "duplicated recipient" error (which the UI would
show) and proceed to remove the recipient too (for unclear reasons).

The fix consists in comparing the "existing" (ie in-rust-memory) and "updated"
(ie. coming from UI) recipients by address (vs whole Participant object).

ET-2762