Fix switch account failure due to bad userId mapping
In rust lib v0.44.0 (~feb 2025) the method `MailUserSession.userId` was changed
to return a `MailUserSessionUserIdResult` instead of the userId as String.
Due to the direct call to `toString` this started failing, as the emitted data
isn't a valid userId anymore. Login still worked but switching account didn't.

This commit passes the userId we get from the loginFlow down to onLoggedIn for
consistency with onTwoPass(userId) and onTwoFa(userId) methods.
No difference compared to getting again the userId from the result of `toUserContext`
was observed.

ET-2826