Fix message body disappearing when interacting with recipients and config changes
The root cause of the issue is that the AndroidView containing the webview held on
to the initial state forever.

This is now an issue in the base recomposition cases, as AndroidView doesn't
re-create at recomposition (as in, runs onUpdate but not the factory "onCreate"
one which was the only one we relied on).
It is an issue in configuration changes (it was missed all along) and a slightly
more subtle one in recipients changes (when typing in any on To, Cc or Bcc) because
the suggestions cause the webview to exit the composition thus being re-created
from stale state when re-entering it.

The solution consists in letting the VM keep the already-existing "displayBodyUiModel"
field of state up-to-date reflecting what changed in the composer. This is done
by effectively rebuilding the template starting from the updated body which we're
already passing along, on the assumtion performance hit of the template building
can be ignore (or at least, are less relevant than the added complexity of having
an additional "displayBody" field being passed along through the js interface).

We also need to ensure that on the webview, the initial value is always set one
time only (through launched effect) and that when we set it it's the correct value
(vs some initial empty value, done through OpenWithMessageAction).

ET-2858