Fix crash when mail settings are updated on a different platform while on mailbox
Due to paging3 lib's requirement of pages being immutable together with the way
the folderColor settings was evaluated (as a combine when mapping each page item)
when it changed it triggered the mapping and recollection of the items without
re-creating the page, causing paging to throw an IllegalStateException.

This solution moves the folder color settings evaluation to being read when a page
is mapped vs observed. This means that it UI will not immediately update when
this setting is changed on another client (a user actions will be needed, eg.
switch label).

ET-2929