Move load embedded image function to composer view model
This is needed due to constraints dictated by the rust's lib architecture,
for which clients need to hold the "Draft" object in memory.
In this case, the object is held by RustDraftDataSource; the issue with introducing
a new view model is that ComposerViewModel and the new MessageBodyEditorViewModel
get two different instances of RustDraftDataSource, where only the one in
ComposerViewModel holds a valid "rust draft" reference.
This in turn, causes any operation on the second instance to fail.

Settings up a custom dagger component and scope it to the comoser screen would
be an alternative solution, which unfortunately I couldn't get to work in an
hour-ish timebox. That'll be further investigated.

ET-2932