Temporary remove schedule-send related events from composer
This is due to rebasing the import of new state management on main, where
part of the schedule send feature was integrated.
This allows to keep PRs small, the feature is adapted on top and being raised
as a separated PR.

This commit could be considered a fixup of the following (split for clearness)
fixup! Fixing compilation failures in view model to adopt new state

ET-3234