Fix "ui freeze" when showing bottomsheet in composer
Due to some flickering issues when hiding the keyboard and showing the bottomsheet,
in some cases the sheet happened to be hidden while the showBottomSheet flag
on composer is true causing the screen to not respond to any inputs anymore.

Similarly as found in ContactListScreen a workaround is adding a small delay
in between the operations.

Also, setting the showBottomSheet flag before changing the bottom sheet state
(in case of show) allows the animation to be visible.

ET-3366