Move all usages of old core ViewMode model to local mail-label ViewMode
The new ViewMode model was added to mail-label module as view mode is strictly related
to the label the current mailbox is in.
Note that the ObserveCurrentViewMode use case is broken by this commit to avoid a
circular dependency, will be fixed in the next ones by moving the impl to rust.

ET-3352