Implement in-memory repository and use cases for selected mail label id
 - This repository and use case classes will replace SelectedMailLabelId class
and provide solution for race conditions causing glitches on the mailbox screen

 - Currently as soon as the user requests to change the location we
emit the updated label and we assume the location has changed.

This causes race conditions since this event is consumed in various
places and causing UI state changes. The screen is updated with the
new location before its message list is actually loaded.

With this refactoring, we will provide separate Flows for selected location (which
means user requested to switch to this location but it's not yet loaded) and  separate flow
for loaded locations

If we want to update the screen immediately without waiting for messages to be
loaded we need to use ObserveSelectedMailLabelId, whereas if we want to update
the screen after messages are loaded (such as showing banners above messages in the new location)
then we need to use ObserveLoadedMailLabelId.

 - To select a new location, selectMailLabelId() should be used

ET-3282